# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulp API",
    "description": "Pulp social-listening HTTP API. This spec is generated from the Rust backend and is the authoritative API contract.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "operationId": "listNotifications",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "operationId": "createNotification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/test": {
      "post": {
        "tags": [
          "notifications"
        ],
        "operationId": "sendTestNotification",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestNotificationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "operationId": "deleteNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "`GET /api/workspaces` — list all workspaces (not workspace-scoped).",
        "operationId": "listWorkspaces",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "`POST /api/workspaces` — create a workspace. Body: `{ name, description? }`.",
        "operationId": "createWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}": {
      "put": {
        "tags": [
          "workspaces"
        ],
        "summary": "`PUT /api/workspaces/:id` — rename/update a workspace. Body: `{ name, description? }`.",
        "operationId": "updateWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "`DELETE /api/workspaces/:id` — delete a workspace (cascades to its monitors and notifications).",
        "operationId": "deleteWorkspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateNotification": {
        "type": "object",
        "description": "Create request for a notification.",
        "required": [
          "workspace_id",
          "kind",
          "config"
        ],
        "properties": {
          "config": {
            "description": "Kind-specific config JSON (webpush: `{endpoint,p256dh,auth}`; webhook:\n`{url}`)."
          },
          "kind": {
            "type": "string",
            "description": "`webpush` | `webhook`."
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "description": "A per-workspace notification: a pure delivery endpoint. Every mention that\nenters the feed fans out to ALL notifications in its workspace — there is no\nmatching, criteria, per-monitor toggle, or frequency. `kind` selects the\ntransport and `config` is the kind-specific JSON:\n- `webpush`: `{ endpoint, p256dh, auth }` (the browser's subscription)\n- `webhook`: `{ url }`",
        "required": [
          "id",
          "workspace_id",
          "kind",
          "config",
          "created_at"
        ],
        "properties": {
          "config": {},
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
      "TestNotificationResult": {
        "type": "object",
        "description": "Result of sending a test notification: how many of the workspace's\nnotifications it attempted to deliver to.",
        "required": [
          "delivered"
        ],
        "properties": {
          "delivered": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "WorkspaceBody": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
...
}